requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
cffi==1.11.5
cryptography==2.4.2
Django==2.0.7
django-phonenumber-field==2.1.0
djangorestframework==3.8.2
idna==2.8
phonenumbers==8.10.1
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.5
six==1.11.0
